refactor: Extract Spring Connectors test
Move the Spring Connectors test into a separate class. Disable the connectors for the other test cases.

Reason: The test is flaky because of an issue with purge and Connectors (see: https://github
.com/camunda/camunda/issues/31542).
