+++ testing/camunda-process-test-spring/src/test/resources/application.yml
@@ -1,11 +1,3 @@
-io:
-  camunda:
-    process:
-      test:
-        connectors-enabled: true
-        connectors-secrets:
-          CONNECTORS_URL: http://connectors:8080/actuator/health/readiness
-
 logging:
   level:
     root: info
